Introduce ems class for ems connection
